Fix DID-based weights storage, model filter, and score calculation
Database changes:
- Change users table: DID as PRIMARY KEY (was handle)
- Change weights table: foreign key references users(did)
- Add GetWeightsByDID() and GetPostsNeedingAnalysis() functions
- Fix underscore keys for Schwartz values in GetPostsWithAnalysisAveraged

Feed service:
- Remove GlobalUserWeights shared memory cache
- Load weights directly from SQLite database
- Fix model filter in GetPostsWithAnalysisAveraged (LIKE '%model%')
- Change Score from int to float64 for weight precision
- Add logging for debugging feed generation

Feed generator:
- Fix model filter flags (must be before subcommand)
- Fix prompt keys to use underscores (matching database columns)
- Add JSON logging for analysis results
- Limit analyses to max 5 per model per post

Breaking changes:
- SaveUser signature: now requires both handle and DID
- SaveWeights/SaveWeight: use DID instead of handle

+++ pkg/schwartz/types.go
@@ -15,7 +15,7 @@ type AIStats struct {
 type ValueAnalysis struct {
 	Rating    SchwartzValues `json:"Rating"`
 	Reasoning string         `json:"Reasoning"`
-	Score     int            `json:"Score"`
+	Score     float64        `json:"Score"`
 	Stats     AIStats        `json:"Stats"`
 	Model     string         `json:"Model,omitempty"`
 	Provider  string         `json:"Provider,omitempty"`
